Updated for Blender 2.8
Updated script for Blender 2.80. Used mesh.polygons in place of mesh.tessfaces in ExportMQO. Removed unused parameters and unused code in ExportMQO. Added filename property in ExportMQO.  Added code to handle ngons but it is not used since loop_triangle UVs are not suitable. Corrected UVs exported to mqo so textured correctly in Metasequoia.